Remove show view for teams, add URL builder
Created a moduel called `STATSURLBuilder`. It's responsible for building
URLs which conform to STATS. Initial tests for this module were written
as well.

The show action was removed from "teams" because to "show" a team is
meaningless. Instead, the user will want the team's ranking within their
division or statistics on an individual player.