Clean up code, completed NHL league standings parser
Although I have yet to thoroughly test the parser, I could see from
printing the output to the screen that the output is what we expect.
That is, an array of hashes in which the has has the keys `name`, which
is the team's name and `stats` which are the team's statistics.