Add a test for the roster scraper.
Changed the return type of `NHLScraper` from a file to text. It did not
make sense to have the intermediary object, which will be the model,
convert a file to text. The responsibility of the scraper is to pull
data from HTML, although the data does not have to be clean, converting
it from a partial HTML document to text is a trivial task.

The raw text will be returned to the model and the model will then pass
it on to the parser for further processing. The separation of concerns
is quite clear.