Remove the expected result
The intention of this file was to match it string for string against the
test subject. I later decided this was extreme overkill. Testing the
subject for nil and for the expected item count is sufficient for this
app. Although ensuring accuracy is ideal, it's also time consuming and
painstaking. I don't believe there is much benefit to verifying the code
to that degree.