Add scraper for NHL team schedule
Also fix bug in NHl team schedule download. There was a for loop which
iterated between the integers 4 and 9 inclusive. What this actually did
was download the schedules of the months in which the NHL was either in
the post season or off completely. This was fixed by defining an array
of month numbers in which the NHL season is active.