Fix bug in schedule parsing
The schedule parser can only parse 1 month at a time, but the original
design allowed for all 12 months. This was fixed so that the parser now
returns a hash of month name to array of games

`{april: [{opponent: 'vs ottowa', date: 'tue 2', tv: 'NESN'}]}`

Fixed the downloader test assertion to look for the correct number of
months in a schedule and not all 12.